requirements.txt
requests>=2.31.0
urllib3>=2.2.0
python-dateutil>=2.8.2
pandas>=2.2.0
sqlalchemy<2.0,>=1.4.32
geopy>=2.4.1
beautifulsoup4>=4.12.3
lxml>=5.2.1
httpx>=0.27.0  # optional for async future work
pydantic>=2.6.0
redis>=4.5.5,<5.0
celery[redis]>=5.3.6
prometheus_client>=0.20.0
flake8>=7.0.0
coverage>=7.5.0
mlflow>=2.12.1
xgboost>=2.0.3
scikit-learn>=1.5.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
tabulate>=0.9.0  # pretty-printing DataFrames in CLI/scripts
streamlit>=1.35.0
lightgbm>=4.3.0
optuna>=3.6.1
feast[redis]>=0.44.1
sentence-transformers>=2.6.1
h3>=3.7.6
pendulum>=2.1.2
evidently>=0.4.18
torch>=2.0
hvac>=1.2.1
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-exporter-jaeger>=1.18.0
opentelemetry-exporter-otlp>=1.18.0
shap>=0.43.0
datahub>=0.12.0
scipy>=1.11.0
web3>=6.12.0
pillow>=10.2.0
requests-oauthlib>=1.3.1
email-validator>=2.0.0
pytest>=8.2.0
great_expectations>=0.18.12
catboost>=1.2.5
pytorch-tabnet>=4.0.0
geopandas>=0.14.0
osmnx>=1.8.0
